git-svn-id: svn://svn.code.sf.net/p/iup/iup/trunk@3937 2f08bc20-984a-4df8-b195-9b5d5d477537

+++ iup/html/examples/tutorial/simple_paint2/simple_paint_file.cpp
@@ -106,7 +106,7 @@ bool SimplePaintFile::SaveAsFile(const char* new_filename)
 
 static void image_fill_white(imImage* image)
 {
-  float color[3];
+  double color[3];
 
   color[0] = 255;
   color[1] = 255;
